refactor:  create folder to shared types
