+++ types/ImageType.tsx
@@ -1,0 +1,8 @@
+import { StaticImageData } from "next/image";
+
+export type ImageType = {
+    alt: string;
+    url: string | StaticImageData;
+    width: number;
+    height: number;
+};
