+++ types/TagType.tsx
@@ -1,0 +1,3 @@
+export type TagType = {
+    name: string;
+};
